Cleaned up the protocol and implemented the initialization

+++ src/libs/shifter/piso_shifter.h
@@ -5,6 +5,6 @@
 
 void piso_shifter_init(void);
 
-uint8_t piso_shifter_get(void);
+uint16_t piso_shifter_get(void);
 
 #endif /* _PISO_SHIFTER_HEADER_ */
